Move `MultiModalConfig` from `config/__init__.py` to `config/multimodal.py`
